Try to make the new memory management more swap friendly
Add swapcache into the other_file; this will let the low memory killer fill all ram; but not try to need swap.

The hope is swap will be filled as cache is needed; but we will not require much swap for the application loading now; thus will not delay the user action.

In addition removed references to other_free as the calculation needed extra CPU and is only used in a debug message at this point in time.